requirements.txt
torch>=2.0.0
numpy>=1.21.0
matplotlib>=3.5.0
plotly>=5.10.0
pandas>=1.3.0
scipy>=1.7.0
xarray>=0.20.0
netCDF4>=1.5.8
cartopy>=0.20.0
tqdm>=4.62.0
pyyaml>=6.0
scikit-learn>=1.0.0
